Add files via upload
This is main code for the term final project of the marching learning class. The selected project uses data (features) that are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass. The data is then used to train a designed neural network to help diagnose whether the suspected cancerous cells are malignant or benign.